moved 'main' packages that didn't need to be in 'main' into their own packages
