+++ traffic_ops/traffic_ops_golang/routing/wrappers_test.go
@@ -1,4 +1,4 @@
-package main
+package routing
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
