+++ traffic_ops/traffic_ops_golang/urisigning/urisigning_test.go
@@ -1,4 +1,4 @@
-package main
+package urisigning
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
